# go/api/v1/account.http.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// source: api/v1/account.proto

package apipb

import (
	bytes "bytes"
	context "context"
	fmt "fmt"
	grpc "google.golang.org/grpc"
	codes "google.golang.org/grpc/codes"
	status "google.golang.org/grpc/status"
	protojson "google.golang.org/protobuf/encoding/protojson"
	proto "google.golang.org/protobuf/proto"
	io "io"
	ioutil "io/ioutil"
	mime "mime"
	http "net/http"
	strings "strings"
)

// AccountServiceHTTPService is the server API for AccountService service.
type AccountServiceHTTPService interface {
	AuthenticateEmail(context.Context, *AuthenticateEmailRequest) (*Session, error)
	RefreshToken(context.Context, *RefreshTokenRequest) (*Session, error)
}

// AccountServiceHTTPConverter has a function to convert AccountServiceHTTPService interface to http.HandlerFunc.
type AccountServiceHTTPConverter struct {
	srv AccountServiceHTTPService
}

// NewAccountServiceHTTPConverter returns AccountServiceHTTPConverter.
func NewAccountServiceHTTPConverter(srv AccountServiceHTTPService) *AccountServiceHTTPConverter {
	return &AccountServiceHTTPConverter{
		srv: srv,
	}
}

// AuthenticateEmail returns AccountServiceHTTPService interface's AuthenticateEmail converted to http.HandlerFunc.
func (h *AccountServiceHTTPConverter) AuthenticateEmail(onError func(ctx context.Context, w http.ResponseWriter, r *http.Request, arg, ret proto.Message, err error), interceptors ...grpc.UnaryServerInterceptor) http.HandlerFunc {
	if onError == nil {
		onError = func(ctx context.Context, w http.ResponseWriter, r *http.Request, arg, ret proto.Message, err error) {
			if err != nil {
				w.WriteHeader(http.StatusInternalServerError)
				p := status.New(codes.Unknown, err.Error()).Proto()
				switch contentType, _, _ := mime.ParseMediaType(r.Header.Get("Content-Type")); contentType {
				case "application/protobuf", "application/x-protobuf":
					buf, err := proto.Marshal(p)
					if err != nil {
						return
					}
					if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
						return
					}
				case "application/json":
					options := protojson.MarshalOptions{UseProtoNames: true}
					buf, err := options.Marshal(p)
					if err != nil {
						return
					}
					if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
						return
					}
				default:
				}
			}
		}
	}
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := r.Context()

		contentType, _, _ := mime.ParseMediaType(r.Header.Get("Content-Type"))

		accepts := strings.Split(r.Header.Get("Accept"), ",")
		accept := accepts[0]
		if accept == "*/*" || accept == "" {
			if contentType != "" {
				accept = contentType
			} else {
				accept = "application/json"
			}
		}

		w.Header().Set("Content-Type", accept)

		arg := &AuthenticateEmailRequest{}
		if r.Method != http.MethodGet {
			body, err := ioutil.ReadAll(r.Body)
			if err != nil {
				onError(ctx, w, r, nil, nil, err)
				return
			}

			switch contentType {
			case "application/protobuf", "application/x-protobuf":
				if err := proto.Unmarshal(body, arg); err != nil {
					onError(ctx, w, r, nil, nil, err)
					return
				}
			case "application/json":
				if err := protojson.Unmarshal(body, arg); err != nil {
					onError(ctx, w, r, nil, nil, err)
					return
				}
			default:
				w.WriteHeader(http.StatusUnsupportedMediaType)
				_, err := fmt.Fprintf(w, "Unsupported Content-Type: %s", contentType)
				onError(ctx, w, r, nil, nil, err)
				return
			}
		}

		n := len(interceptors)
		chained := func(ctx context.Context, arg interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
			chainer := func(currentInter grpc.UnaryServerInterceptor, currentHandler grpc.UnaryHandler) grpc.UnaryHandler {
				return func(currentCtx context.Context, currentReq interface{}) (interface{}, error) {
					return currentInter(currentCtx, currentReq, info, currentHandler)
				}
			}

			chainedHandler := handler
			for i := n - 1; i >= 0; i-- {
				chainedHandler = chainer(interceptors[i], chainedHandler)
			}
			return chainedHandler(ctx, arg)
		}

		info := &grpc.UnaryServerInfo{
			Server:     h.srv,
			FullMethod: "/appsets.api.v1.AccountService/AuthenticateEmail",
		}

		handler := func(c context.Context, req interface{}) (interface{}, error) {
			return h.srv.AuthenticateEmail(c, req.(*AuthenticateEmailRequest))
		}

		iret, err := chained(ctx, arg, info, handler)
		if err != nil {
			onError(ctx, w, r, arg, nil, err)
			return
		}

		ret, ok := iret.(*Session)
		if !ok {
			onError(ctx, w, r, arg, nil, fmt.Errorf("/appsets.api.v1.AccountService/AuthenticateEmail: interceptors have not return Session"))
			return
		}

		switch accept {
		case "application/protobuf", "application/x-protobuf":
			buf, err := proto.Marshal(ret)
			if err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
			if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
		case "application/json":
			options := protojson.MarshalOptions{UseProtoNames: true}
			buf, err := options.Marshal(ret)
			if err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
			if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
		default:
			w.WriteHeader(http.StatusUnsupportedMediaType)
			_, err := fmt.Fprintf(w, "Unsupported Accept: %s", accept)
			onError(ctx, w, r, arg, ret, err)
			return
		}
		onError(ctx, w, r, arg, ret, nil)
	})
}

// AuthenticateEmailWithName returns Service name, Method name and AccountServiceHTTPService interface's AuthenticateEmail converted to http.HandlerFunc.
func (h *AccountServiceHTTPConverter) AuthenticateEmailWithName(onError func(ctx context.Context, w http.ResponseWriter, r *http.Request, arg, ret proto.Message, err error), interceptors ...grpc.UnaryServerInterceptor) (string, string, http.HandlerFunc) {
	return "AccountService", "AuthenticateEmail", h.AuthenticateEmail(onError, interceptors...)
}

// AuthenticateEmailHTTPRule returns HTTP method, path and AccountServiceHTTPService interface's AuthenticateEmail converted to http.HandlerFunc.
func (h *AccountServiceHTTPConverter) AuthenticateEmailHTTPRule(onError func(ctx context.Context, w http.ResponseWriter, r *http.Request, arg, ret proto.Message, err error), interceptors ...grpc.UnaryServerInterceptor) (string, string, http.HandlerFunc) {
	if onError == nil {
		onError = func(ctx context.Context, w http.ResponseWriter, r *http.Request, arg, ret proto.Message, err error) {
			if err != nil {
				w.WriteHeader(http.StatusInternalServerError)
				p := status.New(codes.Unknown, err.Error()).Proto()
				switch contentType, _, _ := mime.ParseMediaType(r.Header.Get("Content-Type")); contentType {
				case "application/protobuf", "application/x-protobuf":
					buf, err := proto.Marshal(p)
					if err != nil {
						return
					}
					if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
						return
					}
				case "application/json":
					options := protojson.MarshalOptions{UseProtoNames: true}
					buf, err := options.Marshal(p)
					if err != nil {
						return
					}
					if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
						return
					}
				default:
				}
			}
		}
	}
	return http.MethodPost, "/api/v1/authenticate/email", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := r.Context()

		contentType, _, _ := mime.ParseMediaType(r.Header.Get("Content-Type"))

		accepts := strings.Split(r.Header.Get("Accept"), ",")
		accept := accepts[0]
		if accept == "*/*" || accept == "" {
			if contentType != "" {
				accept = contentType
			} else {
				accept = "application/json"
			}
		}

		w.Header().Set("Content-Type", accept)

		arg := &AuthenticateEmailRequest{}
		if r.Method != http.MethodGet {
			body, err := ioutil.ReadAll(r.Body)
			if err != nil {
				onError(ctx, w, r, nil, nil, err)
				return
			}

			switch contentType {
			case "application/protobuf", "application/x-protobuf":
				if err := proto.Unmarshal(body, arg); err != nil {
					onError(ctx, w, r, nil, nil, err)
					return
				}
			case "application/json":
				if err := protojson.Unmarshal(body, arg); err != nil {
					onError(ctx, w, r, nil, nil, err)
					return
				}
			default:
				w.WriteHeader(http.StatusUnsupportedMediaType)
				_, err := fmt.Fprintf(w, "Unsupported Content-Type: %s", contentType)
				onError(ctx, w, r, nil, nil, err)
				return
			}
		}

		n := len(interceptors)
		chained := func(ctx context.Context, arg interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
			chainer := func(currentInter grpc.UnaryServerInterceptor, currentHandler grpc.UnaryHandler) grpc.UnaryHandler {
				return func(currentCtx context.Context, currentReq interface{}) (interface{}, error) {
					return currentInter(currentCtx, currentReq, info, currentHandler)
				}
			}

			chainedHandler := handler
			for i := n - 1; i >= 0; i-- {
				chainedHandler = chainer(interceptors[i], chainedHandler)
			}
			return chainedHandler(ctx, arg)
		}

		info := &grpc.UnaryServerInfo{
			Server:     h.srv,
			FullMethod: "/appsets.api.v1.AccountService/AuthenticateEmail",
		}

		handler := func(c context.Context, req interface{}) (interface{}, error) {
			return h.srv.AuthenticateEmail(c, req.(*AuthenticateEmailRequest))
		}

		iret, err := chained(ctx, arg, info, handler)
		if err != nil {
			onError(ctx, w, r, arg, nil, err)
			return
		}

		ret, ok := iret.(*Session)
		if !ok {
			onError(ctx, w, r, arg, nil, fmt.Errorf("/appsets.api.v1.AccountService/AuthenticateEmail: interceptors have not return Session"))
			return
		}

		switch accept {
		case "application/protobuf", "application/x-protobuf":
			buf, err := proto.Marshal(ret)
			if err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
			if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
		case "application/json":
			options := protojson.MarshalOptions{UseProtoNames: true}
			buf, err := options.Marshal(ret)
			if err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
			if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
		default:
			w.WriteHeader(http.StatusUnsupportedMediaType)
			_, err := fmt.Fprintf(w, "Unsupported Accept: %s", accept)
			onError(ctx, w, r, arg, ret, err)
			return
		}
		onError(ctx, w, r, arg, ret, nil)
	})
}

// RefreshToken returns AccountServiceHTTPService interface's RefreshToken converted to http.HandlerFunc.
func (h *AccountServiceHTTPConverter) RefreshToken(onError func(ctx context.Context, w http.ResponseWriter, r *http.Request, arg, ret proto.Message, err error), interceptors ...grpc.UnaryServerInterceptor) http.HandlerFunc {
	if onError == nil {
		onError = func(ctx context.Context, w http.ResponseWriter, r *http.Request, arg, ret proto.Message, err error) {
			if err != nil {
				w.WriteHeader(http.StatusInternalServerError)
				p := status.New(codes.Unknown, err.Error()).Proto()
				switch contentType, _, _ := mime.ParseMediaType(r.Header.Get("Content-Type")); contentType {
				case "application/protobuf", "application/x-protobuf":
					buf, err := proto.Marshal(p)
					if err != nil {
						return
					}
					if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
						return
					}
				case "application/json":
					options := protojson.MarshalOptions{UseProtoNames: true}
					buf, err := options.Marshal(p)
					if err != nil {
						return
					}
					if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
						return
					}
				default:
				}
			}
		}
	}
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := r.Context()

		contentType, _, _ := mime.ParseMediaType(r.Header.Get("Content-Type"))

		accepts := strings.Split(r.Header.Get("Accept"), ",")
		accept := accepts[0]
		if accept == "*/*" || accept == "" {
			if contentType != "" {
				accept = contentType
			} else {
				accept = "application/json"
			}
		}

		w.Header().Set("Content-Type", accept)

		arg := &RefreshTokenRequest{}
		if r.Method != http.MethodGet {
			body, err := ioutil.ReadAll(r.Body)
			if err != nil {
				onError(ctx, w, r, nil, nil, err)
				return
			}

			switch contentType {
			case "application/protobuf", "application/x-protobuf":
				if err := proto.Unmarshal(body, arg); err != nil {
					onError(ctx, w, r, nil, nil, err)
					return
				}
			case "application/json":
				if err := protojson.Unmarshal(body, arg); err != nil {
					onError(ctx, w, r, nil, nil, err)
					return
				}
			default:
				w.WriteHeader(http.StatusUnsupportedMediaType)
				_, err := fmt.Fprintf(w, "Unsupported Content-Type: %s", contentType)
				onError(ctx, w, r, nil, nil, err)
				return
			}
		}

		n := len(interceptors)
		chained := func(ctx context.Context, arg interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
			chainer := func(currentInter grpc.UnaryServerInterceptor, currentHandler grpc.UnaryHandler) grpc.UnaryHandler {
				return func(currentCtx context.Context, currentReq interface{}) (interface{}, error) {
					return currentInter(currentCtx, currentReq, info, currentHandler)
				}
			}

			chainedHandler := handler
			for i := n - 1; i >= 0; i-- {
				chainedHandler = chainer(interceptors[i], chainedHandler)
			}
			return chainedHandler(ctx, arg)
		}

		info := &grpc.UnaryServerInfo{
			Server:     h.srv,
			FullMethod: "/appsets.api.v1.AccountService/RefreshToken",
		}

		handler := func(c context.Context, req interface{}) (interface{}, error) {
			return h.srv.RefreshToken(c, req.(*RefreshTokenRequest))
		}

		iret, err := chained(ctx, arg, info, handler)
		if err != nil {
			onError(ctx, w, r, arg, nil, err)
			return
		}

		ret, ok := iret.(*Session)
		if !ok {
			onError(ctx, w, r, arg, nil, fmt.Errorf("/appsets.api.v1.AccountService/RefreshToken: interceptors have not return Session"))
			return
		}

		switch accept {
		case "application/protobuf", "application/x-protobuf":
			buf, err := proto.Marshal(ret)
			if err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
			if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
		case "application/json":
			options := protojson.MarshalOptions{UseProtoNames: true}
			buf, err := options.Marshal(ret)
			if err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
			if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
		default:
			w.WriteHeader(http.StatusUnsupportedMediaType)
			_, err := fmt.Fprintf(w, "Unsupported Accept: %s", accept)
			onError(ctx, w, r, arg, ret, err)
			return
		}
		onError(ctx, w, r, arg, ret, nil)
	})
}

// RefreshTokenWithName returns Service name, Method name and AccountServiceHTTPService interface's RefreshToken converted to http.HandlerFunc.
func (h *AccountServiceHTTPConverter) RefreshTokenWithName(onError func(ctx context.Context, w http.ResponseWriter, r *http.Request, arg, ret proto.Message, err error), interceptors ...grpc.UnaryServerInterceptor) (string, string, http.HandlerFunc) {
	return "AccountService", "RefreshToken", h.RefreshToken(onError, interceptors...)
}

// RefreshTokenHTTPRule returns HTTP method, path and AccountServiceHTTPService interface's RefreshToken converted to http.HandlerFunc.
func (h *AccountServiceHTTPConverter) RefreshTokenHTTPRule(onError func(ctx context.Context, w http.ResponseWriter, r *http.Request, arg, ret proto.Message, err error), interceptors ...grpc.UnaryServerInterceptor) (string, string, http.HandlerFunc) {
	if onError == nil {
		onError = func(ctx context.Context, w http.ResponseWriter, r *http.Request, arg, ret proto.Message, err error) {
			if err != nil {
				w.WriteHeader(http.StatusInternalServerError)
				p := status.New(codes.Unknown, err.Error()).Proto()
				switch contentType, _, _ := mime.ParseMediaType(r.Header.Get("Content-Type")); contentType {
				case "application/protobuf", "application/x-protobuf":
					buf, err := proto.Marshal(p)
					if err != nil {
						return
					}
					if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
						return
					}
				case "application/json":
					options := protojson.MarshalOptions{UseProtoNames: true}
					buf, err := options.Marshal(p)
					if err != nil {
						return
					}
					if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
						return
					}
				default:
				}
			}
		}
	}
	return http.MethodPost, "/api/v1/authenticate/token/refresh", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := r.Context()

		contentType, _, _ := mime.ParseMediaType(r.Header.Get("Content-Type"))

		accepts := strings.Split(r.Header.Get("Accept"), ",")
		accept := accepts[0]
		if accept == "*/*" || accept == "" {
			if contentType != "" {
				accept = contentType
			} else {
				accept = "application/json"
			}
		}

		w.Header().Set("Content-Type", accept)

		arg := &RefreshTokenRequest{}
		if r.Method != http.MethodGet {
			body, err := ioutil.ReadAll(r.Body)
			if err != nil {
				onError(ctx, w, r, nil, nil, err)
				return
			}

			switch contentType {
			case "application/protobuf", "application/x-protobuf":
				if err := proto.Unmarshal(body, arg); err != nil {
					onError(ctx, w, r, nil, nil, err)
					return
				}
			case "application/json":
				if err := protojson.Unmarshal(body, arg); err != nil {
					onError(ctx, w, r, nil, nil, err)
					return
				}
			default:
				w.WriteHeader(http.StatusUnsupportedMediaType)
				_, err := fmt.Fprintf(w, "Unsupported Content-Type: %s", contentType)
				onError(ctx, w, r, nil, nil, err)
				return
			}
		}

		n := len(interceptors)
		chained := func(ctx context.Context, arg interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
			chainer := func(currentInter grpc.UnaryServerInterceptor, currentHandler grpc.UnaryHandler) grpc.UnaryHandler {
				return func(currentCtx context.Context, currentReq interface{}) (interface{}, error) {
					return currentInter(currentCtx, currentReq, info, currentHandler)
				}
			}

			chainedHandler := handler
			for i := n - 1; i >= 0; i-- {
				chainedHandler = chainer(interceptors[i], chainedHandler)
			}
			return chainedHandler(ctx, arg)
		}

		info := &grpc.UnaryServerInfo{
			Server:     h.srv,
			FullMethod: "/appsets.api.v1.AccountService/RefreshToken",
		}

		handler := func(c context.Context, req interface{}) (interface{}, error) {
			return h.srv.RefreshToken(c, req.(*RefreshTokenRequest))
		}

		iret, err := chained(ctx, arg, info, handler)
		if err != nil {
			onError(ctx, w, r, arg, nil, err)
			return
		}

		ret, ok := iret.(*Session)
		if !ok {
			onError(ctx, w, r, arg, nil, fmt.Errorf("/appsets.api.v1.AccountService/RefreshToken: interceptors have not return Session"))
			return
		}

		switch accept {
		case "application/protobuf", "application/x-protobuf":
			buf, err := proto.Marshal(ret)
			if err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
			if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
		case "application/json":
			options := protojson.MarshalOptions{UseProtoNames: true}
			buf, err := options.Marshal(ret)
			if err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
			if _, err := io.Copy(w, bytes.NewBuffer(buf)); err != nil {
				onError(ctx, w, r, arg, ret, err)
				return
			}
		default:
			w.WriteHeader(http.StatusUnsupportedMediaType)
			_, err := fmt.Fprintf(w, "Unsupported Accept: %s", accept)
			onError(ctx, w, r, arg, ret, err)
			return
		}
		onError(ctx, w, r, arg, ret, nil)
	})
}